Bugfix: incorrect display of min=max metrics
As pointed to in #174,
the display of metrics where the local minimum value equals the
local maximum value fails.
This bugfix addresses the issue with the `curValueRange` array being
incorrectly automatically determined, such that the 'wiggle room' around
the data's min/max is always calculcated on the assumption that the delta
of min/max is at least a hardcoded value of "10".
(This minimum might be debatable, a minimum of 0.0001 worked for me too in my tests,
however minimum values below "1" did upset my grid-drawing algorithm.)

Furthermore a long standing oddity, should be also fixed now: When
the minimum value vor a metric was Zero at some point, the coordinate system
would be drawn, starting at an offset of beneath Zero. Which might be desirable
for some real world data, e.g. Power Consumption which for most devices
(should) never go beneath zero.